fix(frontend): Update rollup to manage svg files
The library was no longer able to be built due to the addition of
SVG files. I've added a new plugin to manage this.

I also made `zxcvbn`, the password strength checker an external dependency
for rollup, I'm still debating if it is the right thing to make this
and others `Peer Dependencies` in the package.json file.