feat(frontend): Fix linking library to application
There are times you want to link your library to an application using
it so you can see components in action, for further development. I've
now fixed this, before there was a problem as there were 2 react versions
same verion though, being used, causing an error, see

https://reactjs.org/warnings/invalid-hook-call-warning.html#duplicate-react

for more of an explanation.