fix(frontend): Refactor Wells and Detail can be set to open
The solution for SVG icons worked in storybook and tests, but when
the library was imported into an actual application, they couldn't
be rendered. Now instead of importing an SVG file, I wrap the SVG
elements in a React function component. I feel it is easier for all
SVGs to work this way.

The Details component can now be set to be open via props, this allows
the component to be rendered open programatically.