fix(frontend): Prevent password being displayed
When used in a form, pressing the `ENTER` key invoked the onClick
function of the button, instead of submitting the form. It turns
out that even though we use the `<button />` tag, the enter key affects
it. By specifying every other button, other than the `SUBMIT` one as
`type="button` and the submit one as `type="submit` prevents this from
happening.