fix(frontend): Publish new package on release published
We previously triggered this action when the release is created,
but it doesn't seem to work, since GitHub changed actions to not
trigger on `Draft Releases`. Looking at our releases, none of them
look like drafts, so I'm changing it to when a `release is published`,
hopefully that will trigger it, for now I've been changing to to run
on `push`, just so we can get working.