fix(frontend): Re-add CommonJS support
We had problems using this library in other projects when it came
to testing, the tests just constantly failed. We have hacked around
a bit to try to get this working and this is what we have ended up
with. We have tried it with the app and tests now do pass.