fix(frontend): Make buttons take the minimum size necessary
I previously made it so that buttons took up the full width of their
parent. I now think that was a bad decision and that by default, they
should only take up as much space as necessary. I have now made it like
this and if a button needs to be bigger, then it should have extra css
applied. In practice, I've found I need to add sizing to almost all buttons
whereas I'd like to avoid that as much as possible.