fix(frontend): Refactor Info Button
I refactored the Info button to be just like the Close one, this makes
maintenance easier. I created a new component from the existing info.svg
and removed it.

I also made a small change to the switch button to remove a bit of CSS
and fixed the associated test since that failed after the change.