fix(frontend): Remove unnecessary dependencies
We have a problem with releases, not sure what is going on, but it
seems related to the number of files published. I reduced it to just
one directory of files under the `dist` dir and it works. This one
however appears to have worked some time ago, but I don't know how,
I've not yet seen one work, so I don't really know what is going on.