fix(frontend): Remove clickaway from select
I have removed the `useClickaway` hook from the `select` component.
This was preventing the `select` component from working in Firefox
and probably other browsers.

Ideally we would also have the ability to click outside the select
and for the dropdown bit to be hidden, but we are moving to another
UI library soon and it doesn't seem worth the effort at this time.

IssueID #100