fix(frontend): Fix package building issues
There was an eslint/prettier plugin error when running `yarn build`,
due to running eslint against the code. It seems to have been to do
with clashing versions of eslint being used, so I added the `resolutions`
option to ensure the version of eslint used.

I changed the `babelHelpers` to the default `bundled`, instead of that
recommended for libraries.