fix(frontend): Fix broken tests
Having upgraded material-ui from v4 to v5, the tests broke. This is
due to the tests needing to be wrapped in a `ThemeProvider` now, not
sure why they didn't previously. I introduced a custom `render` function,
similar to the one we use in `orbis`, that does this for every test.