fix(frontend): Update component to use new themes
I have now fixed the theme not being applied by adding the resolutions
for emotion. I got this solution from

    https://github.com/mui-org/material-ui/issues/24282

It isn[t a fix though, more a temporary resolution, we should keep an
eye on this issue, see what happens or if there is a better option than
what I've done here.