fix(frontend): Fix eslint issues starting storybook
I have moved the export of the Portal, as it doesn't seem to be
exported from `@mui/material`.

I saved `menu-item.component.js` and it did something, but when I
look at the `git diff`, I see no change. However, this does seem to
have fixed it, so maybe it is some hidden character, I don't know,
but it works, so that is good enough for me.