create generic characters counter in the astrosat-ui library
I added a new feature to the textField component in the Astrosat-Ui. We can pass the maxLength property with a number value to the textField component. It will activate the limit characters counter, which disables typing process when the counter reaches the 0 value. When the counter value is 0, the helperText font is in red. The 'Characters max length' text is hardcoded.

IssueID ORB-1225