fix the button component for a new version of Astrosat-ui
I replaced the root CSS class with & character - that fixed the problem with the proper button size when we want to see it in a browser for the first time (we don't need to refresh the page to fix it). I moved the CSS rules (containedSizeLarge, textSizeLarge, outlinedSizeLarge) to the core.js file to fix the large button size for the button size story (I had an issue where a large button has the same size as a small size button).
I added some other small CSS changes to fix all issues with the disabled button. The group button component is broken, but my changes to the button component didn't affect to make this broken.

IssueID: 1137