fix button issues for a new version of Astrosat-ui library
I rebuilt the structure of the button component, so it's easier to read the code. I added text and action with proper values to the baseTheme pallete. I did it because, for example, the old version of coreTheme.palette.action.disabled no longer exists. I added some css for a small version of the button to fix the issue with button size. To fix the problem with padding for iconButton I used edgeStart props to added custom css, which fixed the padding problem. The disadvantage of this solution is that the first iconButton needs to have edge props with the start value (see storybook).

IssueId: 1137