rebuild the button group component for a new version of Astrosat-ui
I overwrote some CSS rules to make the same button group component as we have in the old version of Astrosat-ui. I added the groupedContainedInherit CSS rule to overwrite the background colour for those buttons in the storybook with the original light-grey background colour (color =default). The Material-ui documentation doesn't say anything about the groupContainedInherit CSS rule name. I wanted to see what happens if I add the Inherit word to the groupContained CSS rule. So I created the groupContainedInherit CSS rule, and I was able to change the baton background colour for any value. The same trick can be done with 'warning' and 'info' values.

IssueID: 1137