fix(frontend): tidy up the button group component code
I put the defaultProps back in the core.js file, and I remove some unnecessary props in stories because of the default values. Unfortunately, I couldn't refactor CSS in Mark's suggested way, but I started to note all pieces of code that need to be refactored in the future when I get more experience with material-ui CSS. I fixed all aria-label text.

IssueID: 1137