fix(frontend): fix the menu, menu-item and paper components in the new Astrosat-ui
The name of my branch can be confusing. I created this branch for the input component by mistake. I rebuilt the menu, menu items and paper components. In the palette.js file, I added a new colour for the paper background, which needs to be added to the dark.js file to update the elevation in the Mui-paper CSS global class. It makes the paper colour and box-shadow exactly the same as we have in the old version of the library. There is the link which can explain that: https://mui.com/material-ui/react-paper/\#:\~:text\=The%20change%20of,\). The first element on the menu list has a darker grey background colour than the same element in the old version of Astrosat-ui. Is it acceptable? I have some problems changing that colour.

IssueId: ORB-1137