fix(frontend): fix the Dialog component for the new version of Astrosat-ui
I imported the button and dilaogTitle component from Astrosat-ui library, and it makes more sense for me because we already have those components, so it fixed the button and the dialog title issues. I overrode the AppBar CSS to fix colour issues for this component. There is the MuiDialogContent global CSS class with padding values. For some reason, the top padding doesn't work in this class, so I added inline styling for each DialogContent component to fix it.

IsseID: 0RB-1137