fix(frontend): fix the Select component
The new version of the select component doesn't look the same as the old version, but it is the best I can do with it. I added the shape variable to the palette.js file. I added style to each MenuItem in the select.stories file cause I couldn't find a proper way to override marginLeft and marginRight in the MenuItem component.

IssueId: ORB-1137