fix(frontend): refactor the Select component code
I removed the shape variable from the palette.js file, and I added the value 2 for shape - borderRadius in the core.js file. The default value was too high, making the radius really big. I removed the styling with margins from the stories but to make margin with value zero, I had to rid of marginLeft and marginRight in the menuItem component, so The menuItem component doesn't have margins now. Not sure this is acceptable. In the storybook for the Select component, I added single styling for the backgroundColor (line 53). By default, the first MenuItem has a dark yellow backgroundColor that I have no idea how to fix.

IssueId: ORB-1137