fix(frontend): remove hard-coded color value
I tried to pass the theme to the styled utility following material-ui documentation, but unfortunately, I was unsuccessful. When I try to access colour using theme I have an error: property palette does not exist on the type default theme. I am not sure, but maybe there is something wrong with the theme-provider.js file according to the documentation (https://mui.com/material-ui/migration/migration-v4/#types-property-quot-palette-quot-quot-spacing-quot-does-not-exist-on-type-defaulttheme:~:text=The%20root%20of%20your%20application%20should%20look%20something%20like%20this%3A). I was not abble to solve this problem so I refer to the baseTheme to get the colour value.

IssueId: ORB-1137