fix(frontend): fix the text field component
In the core.js file, I changed the colour value. In the text field component, I added top padding using inline styling; for some reason, using sx props breaks the borderColor. I added the Mui-disabled rule to the input component to fix the colour value. I noticed that the bottom radius should be the same as the top ones - fix that by changing it to the proper value.

IssueId: ORB-11137